backend/requirement.txt
fastapi
uvicorn
torch
transformers
ffmpeg-python
pydantic
python-multipart
librosa
soundfile
opencv-python
tqdm
numpy
sentencepiece
accelerate

openai-whisper
gfpgan
realesrgan